code cleanup, pixi config update and minor code refactor
